requirements.txt
django-taggit
Pillow